app/fastapi/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
ultralytics==8.0.141
numpy==1.25.0
pillow==10.0.0

